+++ tests/testthat/test-001-add.interval.col.R
@@ -32,6 +32,23 @@ test_that("add.interval.col", {
   expect_error(add.interval.col(name="a", FUN="this function does not exist", datatype="interval", desc="test addition"),
                regexp="The function named '.*' is not defined.  Please define the function before calling add.interval.col.",
                info="interval column function must exist (or be NA)")
+
+  # formalsmap
+  expect_error(add.interval.col(name="a", FUN="mean", formalsmap=NA),
+               regexp="formalsmap must be a list",
+               info="formalsmap must be a list")
+  expect_error(add.interval.col(name="a", FUN="mean", formalsmap=list(1)),
+               regexp="formalsmap must be a named list",
+               info="formalsmap must be a named list")
+  expect_error(add.interval.col(name="a", FUN=NA, formalsmap=list(A="b")),
+               regexp="formalsmap may not be given when FUN is NA",
+               info="formalsmap cannot be used with FUN=NA")
+  expect_error(add.interval.col(name="a", FUN="mean", formalsmap=list(A="a", "b")),
+               regexp="All formalsmap elements must be named",
+               info="Each element of formalsmap must be named")
+  expect_error(add.interval.col(name="a", FUN="mean", formalsmap=list(y="a")),
+               regexp="All names for the formalsmap list must be arguments to the function",
+               info="formalsmap arguments must map to function arguments")
   
   expect_equal(
     {
@@ -41,6 +58,7 @@ test_that("add.interval.col", {
     list(FUN=NA,
          values=c(FALSE, TRUE),
          desc="test addition",
+         formalsmap=list(),
          depends=c(),
          datatype="interval"),
     info="interval column assignment works with FUN=NA")
@@ -52,9 +70,23 @@ test_that("add.interval.col", {
     list(FUN="mean",
          values=c(FALSE, TRUE),
          desc="test addition",
+         formalsmap=list(),
          depends=c(),
          datatype="interval"),
     info="interval column assignment works with FUN=a character string")
+  expect_equal(
+    {
+      add.interval.col(name="a", FUN="mean", formalsmap=list(x="values"), desc="test addition")
+      get("interval.cols", PKNCA:::.PKNCAEnv)[["a"]]
+    },
+    list(FUN="mean",
+         values=c(FALSE, TRUE),
+         desc="test addition",
+         formalsmap=list(x="values"),
+         depends=c(),
+         datatype="interval"),
+    info="interval column assignment works with FUN=NA")
+  
 })
 
 # Reset the original state